added route for category in the home page
